pyproject.toml
[build-system]
requires = ["setuptools>=78.1.0", "wheel>=0.46.0"]
build-backend = "setuptools.build_meta"

[project]
name = "glmocr"
version = "0.1.4"
description = "GLM OCR - Optical Character Recognition powered by GLM"
authors = [
  {name = "zai-org", email = "wenmeng.yu@aminer.cn"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["ocr", "glm", "ai", "vision"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

dependencies = [
  "pillow>=12.1.0",
  "numpy>=1.26.0",
  "requests>=2.32.0",
  "pydantic>=2.12.0",
  "PyYAML>=6.0.0",
  "portalocker>=3.2.0",
  "python-dotenv>=1.2.0",
  "tqdm>=4.62.0",
  "pymupdf>=1.24.0",
]

[project.optional-dependencies]
layout = [
  "torch>=2.10.0",
  "torchvision>=0.25.0",
  "transformers>=5.3.0",
  "sentencepiece>=0.2.0",
  "accelerate>=1.13.0",
  "opencv-python-headless>=4.8.0",
]

server = [
  "flask>=3.1.0",
]

selfhosted = [
  "torch>=2.10.0",
  "torchvision>=0.25.0",
  "transformers>=5.3.0",
  "sentencepiece>=0.2.0",
  "accelerate>=1.13.0",
  "pypdfium2>=5.6.0",
  "opencv-python-headless>=4.8.0",
]

all = [
  "torch>=2.10.0",
  "torchvision>=0.25.0",
  "transformers>=5.3.0",
  "sentencepiece>=0.2.0",
  "accelerate>=1.13.0",
  "pypdfium2>=5.6.0",
  "flask>=3.1.0",
  "opencv-python-headless>=4.8.0",
]
dev = [
  "pytest>=9.0.0",
  "black>=26.3.0",
  "flake8>=7.0.0",
  "mypy>=1.19.0",
  "pre-commit>=4.5.0",
]

[project.scripts]
glmocr = "glmocr.cli:main"

[project.urls]
Homepage = "https://github.com/zai-org/glm-ocr"
Repository = "https://github.com/zai-org/glm-ocr"
Documentation = "https://github.com/zai-org/glm-ocr#readme"
Issues = "https://github.com/zai-org/glm-ocr/issues"

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["glmocr*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.pytest.ini_options]
testpaths = ["glmocr/tests"]
python_files = ["test_*.py"]
